Add ability to process flat CSV files #126 (#148)
* Add ability to process flat CSV files #126

Instead of a directory tree with “index.csv” files it is possible to specify a sub-tree with a single file named “index-flat.csv“ where each level is described by two columns. Column "X" contains the name and "X-description" contains the description. Columns/levels are processed from left to right, but the order of "description" columns is not important.

* formatting updates

---------

Co-authored-by: Tom Kralidis <tomkralidis@gmail.com>